# docs/swagger.yaml
definitions:
...
  models.PrestasiSiswas:
    properties:
      bidang_prestasi:
        type: string
      created_at:
        type: string
      id:
        type: integer
      jenis_lomba:
        type: string
      peringkat_lomba:
        type: string
      peringkat_prestasi:
        type: string
      siswa_id:
        type: integer
      tahun_meraih_prestasi:
        type: integer
      tingkat_lomba:
        type: string
      tingkat_prestasi:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Siswa:
    properties:
      agama:
        type: string
      alamat:
        type: string
      cita_cita:
        type: string
      created_at:
        type: string
      desa:
        type: string
      email:
        type: string
      hafalan_juz:
        type: string
      hafalan_surat:
        type: string
      hobi:
        type: string
      id:
        type: integer
      jarak_rumah:
        type: string
      jenis_kelamin:
        type: string
      jumlah_saudara:
        type: string
      kabupaten:
        type: string
      kecamatan:
        type: string
      kode_pos:
        type: string
      nama:
        type: string
      nama_ayah:
        type: string
      nama_ibu:
        type: string
      nama_wali:
        type: string
      nik_ayah:
        type: string
      nik_ibu:
        type: string
      nik_siswa:
        type: string
      nik_wali:
        type: string
      nis_lokal:
        type: string
      nisn:
        type: string
      no_hp_ayah:
        type: string
      no_hp_ibu:
        type: string
      no_hp_siswa:
        type: string
      no_hp_wali:
        type: string
      no_kk:
        type: string
      pekerjaan_ayah:
        type: string
      pekerjaan_ibu:
        type: string
      pekerjaan_wali:
        type: string
      pendidikan_ayah:
        type: string
      pendidikan_ibu:
        type: string
      pendidikan_wali:
        type: string
      penghasilan_orangtua:
        type: string
      penghasilan_wali:
        type: string
      provinsi:
        type: string
      sekolah_id:
        type: integer
      tanggal_lahir:
        type: string
      tempat_lahir:
        type: string
      tempat_tinggal:
        type: string
      transportasi:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  siswaController.UpdateInputPrestasiSiswa:
    properties:
      bidang_prestasi:
        type: string
      jenis_lomba:
        type: string
      peringkat_lomba:
        type: string
      peringkat_prestasi:
        type: string
      tahun_meraih_prestasi:
        type: integer
      tingkat_lomba:
        type: string
      tingkat_prestasi:
        type: string
    type: object
  siswaController.UpdateInputSiswa:
    properties:
      agama:
        type: string
      alamat:
        type: string
      cita_cita:
        type: string
      desa:
        type: string
      email:
        type: string
      hafalan_juz:
        type: string
      hafalan_surat:
        type: string
      hobi:
        type: string
      jarak_rumah:
        type: string
      jenis_kelamin:
        type: string
      jumlah_saudara:
        type: string
      kabupaten:
        type: string
      kecamatan:
        type: string
      kode_pos:
        type: string
      nama:
        type: string
      nama_ayah:
        type: string
      nama_ibu:
        type: string
      nama_wali:
        type: string
      nik_ayah:
        type: string
      nik_ibu:
        type: string
      nik_siswa:
        type: string
      nik_wali:
        type: string
      nis_lokal:
        type: string
      nisn:
        type: string
      no_hp_ayah:
        type: string
      no_hp_ibu:
        type: string
      no_hp_siswa:
        type: string
      no_hp_wali:
        type: string
      no_kk:
        type: string
      pekerjaan_ayah:
        type: string
      pekerjaan_ibu:
        type: string
      pekerjaan_wali:
        type: string
      pendidikan_ayah:
        type: string
      pendidikan_ibu:
        type: string
      pendidikan_wali:
        type: string
      penghasilan_orangtua:
        type: string
      penghasilan_wali:
        type: string
      provinsi:
        type: string
      tanggal_lahir:
        type: string
      tempat_lahir:
        type: string
      tempat_tinggal:
        type: string
      transportasi:
        type: string
    required:
    - nama
    - nisn
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /siswa/{id}:
    delete:
      description: Menghapus semua data siswa berdasarkan id. mulai dari akun, biodata,
        asal sekolah, prestasi, beasiswa.
      parameters:
      - description: Siswa id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Menghapus data siswa.
      tags:
      - Sekolah
    get:
      description: Mendapatkan data siswa berdasarkan id.
      parameters:
      - description: Siswa id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Siswa'
      summary: Mendapatkan data siswa.
      tags:
      - Siswa
    patch:
      description: memperbaharui data siswa berdasarkan id.
      parameters:
      - description: Siswa id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Siswa
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/siswaController.UpdateInputSiswa'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Siswa'
      security:
      - BearerToken: []
      summary: Memperbarui data Siswa.
      tags:
      - Siswa
...
  /siswa/{id}/prestasi:
    get:
      description: Mendapatkan data prestasi siswa berdasarkan id siswa.
      parameters:
      - description: Siswa ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PrestasiSiswas'
      summary: Mendapatkan data prestasi siswa berdasarkan id siswa.
      tags:
      - Siswa -> Prestasi
...
  /siswa/prestasi/{id}:
    get:
      description: Mendapatkan data prestasi siswa berdasarkan id.
      parameters:
      - description: Prestasi_siswa ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PrestasiSiswas'
      summary: Mendapatkan data prestasi siswa.
      tags:
      - Siswa -> Prestasi
    patch:
      description: memperbaharui data prestasi siswa berdasarkan id.
      parameters:
      - description: Prestasi_siswa id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update prestasi Siswa
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/siswaController.UpdateInputPrestasiSiswa'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PrestasiSiswas'
      security:
      - BearerToken: []
      summary: Memperbarui data prestasi siswa.
      tags:
      - Siswa -> Prestasi
  /siswa/user/{id}:
    get:
      description: Mendapatkan data siswa berdasarkan user ID.
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Siswa'
      summary: Mendapatkan data siswa berdasarkan user ID.
      tags:
      - Siswa
...
swagger: "2.0"